chore(blockifier): move test struct to test file
